Trying to save personal data but not achieved yet

+++ src/main/java/com/example/demo/entities/UserProfile.java
@@ -32,6 +32,6 @@ public class UserProfile {
 	private int weightKg;
 	private int age;
 	private String gender;
-	private Long dailyStepsGoal;
+	private double dailyStepsGoal;
 
 }
